+++ src/index.rs
@@ -65,6 +65,7 @@
         int_format_into,
         nonzero_from_str_radix,
         result_option_map_or_default,
+        strip_circumfix,
     )
 )]
 #![cfg_attr(all(nightly_stable_1_98, feature = "alloc"), feature(box_as_ptr,))]
@@ -103,7 +104,6 @@
         refcell_try_map,
         str_as_str,
         // str_from_utf16_endian,
-        strip_circumfix,
         substr_range,
         supertrait_item_shadowing,
         unsafe_cell_from_mut,